Impose a more sensible ordering for animation graph evaluation.
This commit changes the animation graph evaluation to be operate in a
more sensible order and updates the semantics of blend nodes to conform
to [the animation composition RFC]. Prior to this patch, a node graph
like this:

                ┌─────┐
                │     │
                │  1  │
                │     │
                └──┬──┘
                   │
           ┌───────┴───────┐
           │               │
           ▼               ▼
        ┌─────┐         ┌─────┐
        │     │         │     │
        │  2  │         │  3  │
        │     │         │     │
        └──┬──┘         └──┬──┘
           │               │
       ┌───┴───┐       ┌───┴───┐
       │       │       │       │
       ▼       ▼       ▼       ▼
    ┌─────┐ ┌─────┐ ┌─────┐ ┌─────┐
    │     │ │     │ │     │ │     │
    │  4  │ │  6  │ │  5  │ │  7  │
    │     │ │     │ │     │ │     │
    └─────┘ └─────┘ └─────┘ └─────┘

Would be evaluated as (((4 ⊕ 5) ⊕ 6) ⊕ 7), with the blend (lerp/slerp)
operation notated as ⊕. As quaternion multiplication isn't commutative,
this is very counterintuitive and will especially lead to trouble with
the forthcoming additive blending feature (#15198).

This patch fixes the issue by changing the evaluation order to
postorder, with children of a node evaluated in ascending order by node
index.

To do so, this patch revamps `Keyframes` to be based on an *evaluation
stack* and a *blend register*. During target evaluation, the graph
evaluator traverses the graph in postorder. When encountering a clip
node, the evaluator pushes the possibly-interpolated value onto the
evaluation stack. When encountering a blend node, the evaluator pops
values off the stack into the blend register, accumulating weights as
appropriate. When the graph is completely evaluated, the top element on
the stack is *committed* to the property of the component.

A new system, the *graph threading* system, is added in order to cache
the sorted postorder traversal to avoid the overhead of sorting children
at animation evaluation time. Mask evaluation has been moved to this
system so that the graph only has to be traversed at most once per
frame. Unlike the `ActiveAnimation` list, the *threaded graph* is cached
from frame to frame and only has to be regenerated when the animation
graph asset changes.

This patch currently regresses the `animate_target` performance in
`many_foxes` by around 50%, resulting in an FPS loss of about 2-3 FPS.
I'd argue that this is an acceptable price to pay for a much more
intuitive system. In the future, we can mitigate the regression with a
fast path that avoids consulting the graph if only one animation is
playing. However, in the interest of keeping this patch simple, I didn't
do so here.

[the animation composition RFC]: https://github.com/bevyengine/rfcs/blob/main/rfcs/51-animation-composition.md

+++ crates/bevy_gltf/src/loader.rs
@@ -313,7 +313,8 @@ async fn load_gltf<'a, 'b, 'c>(
                         ReadOutputs::MorphTargetWeights(weights) => {
                             let weights: Vec<_> = weights.into_f32().collect();
                             Box::new(MorphWeightsKeyframes {
-                                morph_target_count: weights.len() / keyframe_timestamps.len(),
+                                morph_target_count: (weights.len() / keyframe_timestamps.len())
+                                    as u32,
                                 weights,
                             }) as Box<dyn Keyframes>
                         }
